% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that_compiled.R
\name{attempt_to_compile}
\alias{attempt_to_compile}
\title{Attempts to compile EDF interface}
\usage{
attempt_to_compile()
}
\description{
Attempts to compile EDF interface
}
\examples{
attempt_to_compile
}
\keyword{internal}
