% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_consistency_flag.R
\name{check_consistency_flag}
\alias{check_consistency_flag}
\title{Checks consistency flag, stops if invalid, returns code if valid.}
\usage{
check_consistency_flag(consistency)
}
\arguments{
\item{consistency}{consistency check control for the time stamps of the start
and end events, etc. Could be \code{'no consistency check'} (0),
\code{'check consistency and report'} (1), \code{'check consistency and fix'} (2).}
}
\value{
integer index
}
\description{
Checks consistency flag, stops if invalid, returns code if valid.
}
\examples{
check_consistency_flag('no consistency check')
}
\keyword{internal}
