% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logical_flag.R
\name{check_logical_flag}
\alias{check_logical_flag}
\title{Checks for validity of a logical flag, stops if not valid}
\usage{
check_logical_flag(logical_flag)
}
\arguments{
\item{logical_flag}{Logical scalar}
}
\value{
logical, value of the logical_flag
}
\description{
Checks for validity of a logical flag, stops if not valid
}
\examples{
import_samples <- TRUE
check_logical_flag(import_samples)
}
