% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that_compiled.R
\name{check_that_compiled}
\alias{check_that_compiled}
\title{Checks whether EDF API library was present and
interface was successfully be compiled}
\usage{
check_that_compiled(fail_loudly = TRUE)
}
\arguments{
\item{fail_loudly}{logical, whether lack of compiled library means
error (\code{TRUE}), just warning (\code{FALSE}), or silent (\code{NA},
for test use only).}
}
\value{
logical
}
\description{
Checks whether EDF API library was present and
interface was successfully be compiled
}
\examples{
check_that_compiled(fail_loudly = FALSE)
}
