% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compiled_library_status}
\alias{compiled_library_status}
\title{Status of compiled library}
\usage{
compiled_library_status()
}
\value{
logical
}
\description{
Return status of compiled library
}
\examples{
compiled_library_status()
}
