% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_codes.R
\name{convert_header_codes}
\alias{convert_header_codes}
\title{Converts integer constants in trial headers to factor with explicit labels}
\usage{
convert_header_codes(trial_headers)
}
\arguments{
\item{trial_headers}{data.frame that contains trial headers.}
}
\description{
Converts integer constants in trial headers to factor with explicit labels.
Please refer to EDF API manual for further details.
}
\keyword{internal}
