% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_NAs}
\alias{convert_NAs}
\title{Convert -32767 (missing info) to NA}
\usage{
convert_NAs(original_frame)
}
\arguments{
\item{original_frame}{data.frame to be processed}
}
\value{
processed data.frame
}
\description{
Converts all -32767 (smallest INT16  value indicating missing info) to NA.
You don't need to call this function directly, as it is automatically evoked within
\code{\link{read_edf}} function.
}
\examples{
\donttest{
  data(gaze)
  gaze$samples <- convert_NAs(gaze$samples)
}
}
