% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_codes.R
\name{convert_recording_codes}
\alias{convert_recording_codes}
\title{Converts integer constants in recordings to factor with explicit labels}
\usage{
convert_recording_codes(trial_recordings)
}
\arguments{
\item{trial_recordings}{data.frame that contains trial recordings.}
}
\description{
Converts integer constants in trial recordings information to factor with explicit labels.
Please refer to EDF API manual for further details.
}
\keyword{internal}
