% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Check if the library was compiled}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{character}

\item{pkgname}{character}
}
\value{
No return value, called to check and warn if compiled EDF API interface is missing.
}
\description{
Check if the library was compiled
}
\keyword{internal}
