% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Compiles the library linking it to EDF API}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{character}

\item{pkgname}{character}
}
\value{
No return value, called to compile EDF API interface.
}
\description{
Compiles the library linking it to EDF API
}
\keyword{internal}
