% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Imported example.edf, events only}
\format{
An object of class \code{edfRecording} of length 9.
}
\usage{
example
}
\description{
An \code{\link{eyelinkRecording}} object for \emph{example.edf},
events only. See \code{\link{eyelinkRecording}} for details.
}
\keyword{datasets}
