% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_with_samples}
\alias{example_with_samples}
\title{Imported example.edf, events and samples}
\format{
An object of class \code{edfRecording} of length 10.
}
\usage{
example_with_samples
}
\description{
An \code{\link{eyelinkRecording}} object for \emph{example.edf},
events and samples with all attributes.
See \code{\link{eyelinkRecording}} for details.
}
\keyword{datasets}
