% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_specific_events.R
\name{extract_fixations}
\alias{extract_fixations}
\title{Extract fixations}
\usage{
extract_fixations(events)
}
\arguments{
\item{events}{An \code{\link[=edfRecording$events]{events}} table of the \code{\link{edfRecording}} object.}
}
\value{
A data.frame with information on \code{\link[=edfRecording$fixations]{fixations}}
}
\description{
Extracts fixations from the events table. Normally, you don't need to call this function yourself,
as it is envoked during the \code{\link{read_edf}} with default settings (\emph{e.g.}, \code{import_fixations = TRUE}).
}
\examples{
# fixations are extracted during the initial read_edf call
recording <- read_edf(system.file("extdata", "example.edf", package = "edfR"),
                      import_fixations = TRUE)

# fixations are extracted during the initial read_edf call by default
recording <- read_edf(system.file("extdata", "example.edf", package = "edfR"))

# fixations are extracted later
recording <- read_edf(system.file("extdata", "example.edf", package = "edfR"),
                      import_fixations = FALSE)
recording$fixations <- extract_fixations(recording$events)
}
\seealso{
read_edf, edfRecording
}
