% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkReader-package.R
\docType{package}
\name{eyelinkReader-package}
\alias{eyelinkReader}
\alias{eyelinkReader-package}
\title{eyelinkReader: Import Gaze Data for EyeLink Eye Tracker}
\description{
Import gaze data from edf files generated by the SR Research \url{https://www.sr-research.com/} EyeLink eye tracker. Gaze data, both recorded events and samples, is imported per trial. The package allows to extract events of interest, such as saccades, blinks, etc. as well as recorded variables and custom events (areas of interest, triggers) into separate tables. The package requires EDF API library that can be obtained at \url{https://www.sr-support.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexander-pastukhov/eyelinkReader/}
  \item \url{https://alexander-pastukhov.github.io/eyelinkReader/}
  \item Report bugs at \url{https://github.com/alexander-pastukhov/eyelinkReader/issues}
}

}
\author{
\strong{Maintainer}: Alexander Pastukhov \email{pastukhov.alexander@gmail.com} (\href{https://orcid.org/0000-0002-8738-8591}{ORCID})

}
\keyword{internal}
