% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkReader-package.R
\docType{package}
\name{eyelinkReader}
\alias{eyelinkReader}
\alias{eyelinkReader-package}
\title{eyelinkReader: Import Gaze Data for EyeLink Eye Tracker}
\description{
Imports gaze data recorded by a SR Research EyeLink eye tracker from an EDF file.
Includes options to import events and/or recorded samples and extract
individual events such as saccades, fixations, blinks, and recorded variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexander-pastukhov/eyelinkReader/}
  \item \url{https://alexander-pastukhov.github.io/eyelinkReader/}
  \item Report bugs at \url{https://github.com/alexander-pastukhov/eyelinkReader/issues}
}

}
\author{
\strong{Maintainer}: Alexander Pastukhov \email{pastukhov.alexander@gmail.com} (\href{https://orcid.org/0000-0002-8738-8591}{ORCID})

}
