% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaze}
\alias{gaze}
\title{Imported example.edf, events and samples}
\format{
An object of class \code{eyelinkRecording} of length 9.
}
\usage{
gaze
}
\description{
An \code{\link{eyelinkRecording}} object for \emph{example.edf},
events and samples with all attributes.
See \code{\link{eyelinkRecording}} for details.
}
\keyword{datasets}
