% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_that_compiled.R
\name{is_compiled}
\alias{is_compiled}
\title{Determine whether library is compiled and can be used}
\usage{
is_compiled()
}
\value{
logical
}
\description{
Determine whether library is compiled and can be used
}
\examples{
is_compiled()
}
