% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{library_version}
\alias{library_version}
\title{Version of the EDF API library}
\usage{
library_version()
}
\description{
Returns version of the EDF API library used to interface an EDF file.
}
\examples{
edfR::library_version()
}
