% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_index_for_sample_attributes.R
\name{logical_index_for_sample_attributes}
\alias{logical_index_for_sample_attributes}
\title{Checks validity of sample attribute vector and returns logical indexes.}
\usage{
logical_index_for_sample_attributes(import_samples, sample_attributes)
}
\arguments{
\item{import_samples}{logical, whether importing was requested.}

\item{sample_attributes}{character vector with names of attributes.}
}
\value{
logical vector
}
\description{
Checks validity of sample attribute vector and returns logical indexes.
}
\examples{
logical_index_for_sample_attributes(TRUE, c('time', 'gx', 'gy'))
}
\keyword{internal}
