% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eyelinkRecording}
\alias{print.eyelinkRecording}
\title{Print info about \code{\link{eyelinkRecording}}}
\usage{
\method{print}{eyelinkRecording}(x, ...)
}
\arguments{
\item{x}{\code{\link{eyelinkRecording}} object}

\item{...}{Addition parameters (unused)}
}
\value{
No return value, called for printing to console.
}
\description{
Print info about \code{\link{eyelinkRecording}}
}
\examples{
\donttest{
  if (eyelinkReader::compiled_library_status()) {
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))
    print(recording)
  }
}
}
