% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_edf.R
\name{read_edf}
\alias{read_edf}
\title{Read EDF file}
\usage{
read_edf(file, consistency = "check consistency and report",
  import_events = TRUE, import_recordings = TRUE, import_samples = FALSE,
  sample_attributes = NULL, start_marker_string = "TRIALID",
  end_marker_string = "TRIAL OK", screen = list(), convert_codes = TRUE,
  import_saccades = TRUE, import_blinks = TRUE, import_fixations = TRUE,
  import_variables = TRUE)
}
\arguments{
\item{file}{full name of the EDF file}

\item{consistency}{consistency check control (for the time stamps of the start
and end events, etc). Could be \code{'no consistency check'},
\code{'check consistency and report'} (default), \code{'check consistency and fix'}.}

\item{import_variables}{}
}
\description{
Reads EDF file and returns an \code{\link{edfRecording}} object that contains events, samples,
and recordings, as well as specific events such as saccades, fixations, blinks, etc.
}
