% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_preamble}
\alias{read_preamble}
\title{Reads preamble of the EDF file.}
\usage{
read_preamble(filename)
}
\value{
string with the preamble
}
\description{
Reads preamble of the EDF file.
}
\details{
Reads preamble of the EDF file.
}
\examples{
read_preamble(system.file("extdata", "example.edf", package = "edfR"))
}
