% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_preamble_str}
\alias{read_preamble_str}
\title{Reads preamble of the EDF file as a single string.}
\usage{
read_preamble_str(filename)
}
\value{
string with the preamble
}
\description{
Reads preamble of the EDF file as a single string.
Please, do not use this function directly. Instead, call \code{\link{read_preamble}} function
that provides a more consistent interface.
}
\examples{
read_preamble(system.file("extdata", "example.edf", package = "edfR"))
}
\keyword{internal}
