% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_samples_to_monocular.R
\name{simplify_samples_to_monocular}
\alias{simplify_samples_to_monocular}
\alias{simplify_samples_to_monocular.data.frame}
\alias{simplify_samples_to_monocular.eyelinkRecording}
\title{Simplifies samples by averaging over binocular recorded properties}
\usage{
simplify_samples_to_monocular(object, fun = mean)

\method{simplify_samples_to_monocular}{data.frame}(object, fun = mean)

\method{simplify_samples_to_monocular}{eyelinkRecording}(object, fun = mean)
}
\arguments{
\item{object}{Either an \code{\link{eyelinkRecording}} object or data.frame with samples,
i.e., \code{samples} slot of the \code{\link{eyelinkRecording}} object.}

\item{fun}{Function used to average across eyes, defaults to \code{\link{mean}}.}
}
\value{
Object of the same time as input, i.e., either a \code{\link{eyelinkRecording}} object
with \emph{modified} \code{samples} slot or a data.frame with simplified monocular samples.
}
\description{
Simplifies samples by averaging over binocular recorded properties
such as \code{px}, \code{py}, \code{hx}, \code{hy}, \code{pa},
\code{gx}, \code{gy}, etc. Uses function specified via \code{fun}
parameter to compute the average with \code{na.rm = TRUE} option.
In case of a monocular recording or information from one eye missing,
this corresponds to using information from one eye only.
}
\examples{
data(gaze)

# by passing samples table
monocular_samples <- simplify_samples_to_monocular(gaze$samples)

# by passing the recording
gaze <- simplify_samples_to_monocular(gaze)
}
