% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Checks for}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{character}

\item{pkgname}{character}

\item{files}{vector with filenames}

\item{folders}{vector with folders}
}
\value{


character (folder) or NULL (no folder contains all files)
}
\description{
Checks for
}
\examples{
Finds a folder that contains ALL files

}
\keyword{internal}
