% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_specific_events.R
\name{extract_AOIs}
\alias{extract_AOIs}
\title{Extracts rectangular areas of interest}
\usage{
extract_AOIs(events)
}
\arguments{
\item{events}{An \code{events} table of the \code{\link{eyelinkRecording}} object.}
}
\value{
A data.frame with the list of \code{\link{eyelinkRecording}}
}
\description{
Extracts rectangular areas of interest (AOI), as defined by "!V IAREA RECTANGLE" command.
Specifically, we expect it to be in format \code{!V IAREA RECTANGLE <index> <left> <top> <right> <bottom> <label>},
where \code{<label>} is a string label and all other variables are integer.
}
\examples{
if (eyelinkReader::is_compiled()) {
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))
    recording$AOIs <- extract_AOIs(recording$events)
}
}
