% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_specific_events.R
\name{extract_blinks}
\alias{extract_blinks}
\title{Extract blinks}
\usage{
extract_blinks(events)
}
\arguments{
\item{events}{An \code{\link[=eyelinkRecording]{events}} table of the \code{\link{eyelinkRecording}} object.}
}
\value{
A data.frame with information on \code{\link[=eyelinkRecording]{blinks}}
}
\description{
Extracts blinks from the events table. Normally, you don't need to call this function yourself,
as it is called during the \code{\link{read_edf}} with default settings (\emph{e.g.}, \code{import_blinks = TRUE}).
}
\examples{
if (eyelinkReader::is_compiled()) {
    # blinks are extracted during the initial read_edf call
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_blinks = TRUE)

    # blinks are extracted during the initial read_edf call by default
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))

    # blinks are extracted later
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_blinks = FALSE)
    recording$blinks <- extract_blinks(recording$events)
}
}
\seealso{
read_edf, eyelinkRecording
}
