% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_specific_events.R
\name{extract_fixations}
\alias{extract_fixations}
\title{Extract fixations}
\usage{
extract_fixations(events)
}
\arguments{
\item{events}{An \code{\link[=eyelinkRecording]{events}} table of the \code{\link{eyelinkRecording}} object.}
}
\value{
A data.frame with information on \code{\link[=eyelinkRecording]{fixations}}
}
\description{
Extracts fixations from the events table. Normally, you don't need to call this function yourself,
as it is called during the \code{\link{read_edf}} with default settings (\emph{e.g.}, \code{import_fixations = TRUE}).
}
\examples{
if (eyelinkReader::is_compiled()) {
    # fixations are extracted during the initial read_edf call
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_fixations = TRUE)

    # fixations are extracted during the initial read_edf call by default
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))

    # fixations are extracted later
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_fixations = FALSE)
    recording$fixations <- extract_fixations(recording$events)
}
}
\seealso{
read_edf, eyelinkRecording
}
