% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_saccades.R
\name{extract_saccades}
\alias{extract_saccades}
\alias{extract_saccades.data.frame}
\alias{extract_saccades.eyelinkRecording}
\title{Extract saccades from recorded events}
\usage{
extract_saccades(object)

\method{extract_saccades}{data.frame}(object)

\method{extract_saccades}{eyelinkRecording}(object)
}
\arguments{
\item{object}{Either an \code{\link{eyelinkRecording}} object or data.frame with events,
i.e., \code{events} slot of the \code{\link{eyelinkRecording}} object.}
}
\value{
Object of the same time as input, i.e., either a \code{\link{eyelinkRecording}} object
with an additional \code{saccades} slot or a data.frame with saccades' information. See
\code{\link{eyelinkRecording}} for details.
}
\description{
Extract saccades from the \code{events} table. Normally, you don't need to call this function yourself,
as it is called during the \code{\link{read_edf}} with default settings (\emph{e.g.}, \code{import_saccades = TRUE}).
}
\examples{
if (eyelinkReader::is_compiled()) {
    # saccades are extracted during the initial read_edf call
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_saccades = TRUE)

    # saccades are extracted during the initial read_edf call by default
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))

    # saccades are extracted later
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_saccades = FALSE)

    # by passing the recording
    recording <- extract_saccades(recording)

    # by passing events table
    saccades <- extract_saccades(recording$events)
}
}
\seealso{
read_edf, eyelinkRecording
}
