% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_specific_events.R
\name{extract_triggers}
\alias{extract_triggers}
\title{Extract triggers, a custom message type}
\usage{
extract_triggers(events)
}
\arguments{
\item{events}{An \code{\link[=edfRecording$events]{events}} table of the \code{\link{edfRecording}} object.}
}
\value{
A data.frame with information on \code{\link[=edfRecording$triggers]{triggers}}
}
\description{
Extracts trigger events, messages that adhere to a "TRIGGER label" format.
Their purpose is to identify the time instance of specific interest.
Please note that due to a non-standard nature of this function \strong{is not} envoked
during the \code{\link{read_edf}} call and you need to call it separately.
}
\examples{
recording <- read_edf(system.file("extdata", "example.edf", package = "edfR"))
recording$triggers <- extract_triggers(recording$events)
}
\seealso{
read_edf, edfRecording
}
