% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_specific_events.R
\name{extract_variables}
\alias{extract_variables}
\title{Extract variables}
\usage{
extract_variables(events)
}
\arguments{
\item{events}{An \code{\link[=edfRecording$events]{events}} table of the \code{\link{edfRecording}} object.}
}
\value{
A data.frame with information on \code{\link[=edfRecording$variables]{variables}}
}
\description{
Extracts variables from the events table. Normally, you don't need to call this function yourself,
as it is envoked during the \code{\link{read_edf}} with default settings (\emph{e.g.}, \code{import_variables = TRUE}).
}
\examples{
if (eyelinkReader::is_compiled()) {
    # variables are extracted during the initial read_edf call
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_variables = TRUE)

    # variables are extracted during the initial read_edf call by default
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))

    # variables are extracted later
    recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"),
                          import_variables = FALSE)
   recording$variables <- extract_variables(recording$events)
}
}
\seealso{
read_edf, edfRecording
}
