% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{eyelinkReader}
\alias{eyelinkReader}
\title{eyelinkReader}
\description{
Imports gazre data recorded by a SR Research Eyelink eye tracker from an EDF file.
Includes options to import events and/or recorded samples and extract
individual events such as saccades, fixations, blinks, and recorded variables.

Please note that although this package is licensed for use under the MIT software license,
but the following additional conditions apply to header files in src/SRResearch subfolder:
All EyeLink® related files, including compiled files may be made
available to SR Research licensed users only and may not otherwise
be redistributed in any manner.
}
