% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyelinkReader-classes.R
\docType{class}
\name{eyelinkRecording-class}
\alias{eyelinkRecording-class}
\alias{eyelinkRecording}
\title{Class \code{eyelinkRecording}.}
\description{
S3 class containing information imported from an edf-file.
}
\details{
See \code{methods(class = "eyelinkRecording")} for an overview of available methods.
}
\section{Slots}{

\describe{
\item{\code{preamble}}{A preamble of the recording, see also \code{\link{read_preamble}} and \code{\link{eyelinkPreamble}}.}

\item{\code{events}}{Events table which is a collection of all \code{FEVENT} imported from the EDF file. See description below.}

\item{\code{samples}}{Samples table  which is a collection of all \code{FSAMPLE} imported from the EDF file. See description below.}
}}

\section{Events}{

Events table which is a collection of all \code{FEVENT} imported from the EDF file.
Column descriptions were copied directly  from the \emph{EDF access C API manual}.
Please refer to that manual for further details. Additional non-standard fields are marked in bold.
\itemize{
\item \code{trial} Trial index, starts at 1.
\item \code{time} Time of event.
\item \code{type} Event type.
\item \code{read} Flags which items were included.
\item \code{sttime} Start time of the event.
\item \code{entime} End time of the event.
\item \strong{\code{sttime_rel}} Start time of the event, relative to the start time of the trial.
\item \strong{\code{entime_rel}} End time of the event, relative to the start time of the trial.
\item \code{hstx}, \code{hsty} Head reference starting points.
\item \code{gstx}, \code{gsty} Gaze starting points.
\item \code{sta} Pupil size at start.
\item \code{henx}, \code{heny} Headref ending points.
\item \code{genx}, \code{geny} Gaze ending points.
\item \code{ena} Pupil size at end.
\item \code{havx}, \code{havy} Headref averages.
\item \code{gavx}, \code{gavy} Gaze averages.
\item \code{ava} Average pupil size.
\item \code{avel} Accumulated average velocity.
\item \code{pvel} Accumulated peak velocity.
\item \code{svel} Start velocity.
\item \code{evel} End velocity.
\item \code{supd_x}, \code{supd_y} Start units-per-degree.
\item \code{eupd_x}, \code{eupd_y} End units-per-degree.
\item \code{eye} Either \code{'LEFT'} (1) or \code{'RIGHT'} (2).
\item \code{status} Error, warning flags.
\item \code{flags} Flags to indicate contents.
\item \code{input} Extra (input word).
\item \code{buttons} Button state and changes.
\item \code{parsedby} 7 bits of flags, PARSEDBY codes.
\item \code{message} Any message string.
}
}

\section{Samples}{
NA
}

\seealso{
\code{\link{fit_cumhist}}
}
