% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.edfRecording}
\alias{print.edfRecording}
\title{Print info about \code{\link{edfRecording}}}
\usage{
\method{print}{edfRecording}(x, ...)
}
\arguments{
\item{x}{\code{\link{edfRecording}} object}

\item{...}{}
}
\value{

}
\description{
Print info about \code{\link{edfRecording}}
}
\examples{
recording <- read_edf(system.file("extdata", "example.edf", package = "eyelinkReader"))
print(recording)
}
