% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_edf_file}
\alias{read_edf_file}
\title{Internal funciton that reads EDF file}
\usage{
read_edf_file(filename, consistency, import_events, import_recordings,
  import_samples, sample_attr_flag, start_marker_string, end_marker_string,
  pixels_per_degree)
}
\arguments{
\item{std::string}{filename, full name of the EDF file}

\item{int}{consistency, consistency check control (for the time stamps of the start
and end events, etc). 0, no consistency check. 1, check consistency and report.
2, check consistency and fix.}

\item{bool}{import_events, load/skip loading events.}

\item{bool}{import_recordings, load/skip loading recordings.}

\item{bool}{import_samples, load/skip loading of samples.}

\item{LogicalVector}{sample_attr_flag, boolean vector that indicates which sample fields are to be stored}

\item{std::string}{start_marker_string, event that marks trial start. Defaults to "TRIALID", if empty.}

\item{std::string}{end_marker_string, event that marks trial end}

\item{NumericVector}{pixels_per_degree, pixels per degree for the screen that was used for the recording
positive values overwrite rx and ry fields' values from the file}
}
\value{
List, contents of the EDF file. Please see read_edf for details.
}
\description{
Reads EDF file into a list that contains events, samples, and recordings.
DO NOT call this function directly. Instead, use read_edf function that implements
parameter checks and additional postprocessing.
}
\details{
Internal funciton that reads EDF file
}
